backend/ezeyway/requirements.txt
asgiref==3.9.1
Django==5.2.5
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2
Pillow==10.4.0
djangorestframework==3.14.0
django-cors-headers==4.3.1